# Perf: modify openmp strategy in module_gint