--- a/source/module_hamilt_lcao/module_gint/gint_force_cpu_interface.cpp
+++ b/source/module_hamilt_lcao/module_gint/gint_force_cpu_interface.cpp
@@ -28,7 +28,7 @@ void Gint::gint_kernel_force(Gint_inout* inout) {
     std::vector<int> block_index(max_size+1,0);
     std::vector<int> block_size(max_size,0);
     std::vector<double> vldr3(this->bxyz,0.0);
-#pragma omp for
+#pragma omp for schedule(dynamic)
     for (int grid_index = 0; grid_index < this->nbxx; grid_index++) {
         const int na_grid = this->gridt->how_many_atoms[grid_index];
         if (na_grid == 0) {
@@ -153,7 +153,7 @@ void Gint::gint_kernel_force_meta(Gint_inout* inout) {
     std::vector<int> block_size(max_size,0);
     std::vector<double> vldr3(this->bxyz,0.0);
     std::vector<double> vkdr3(this->bxyz,0.0);
-#pragma omp for
+#pragma omp for schedule(dynamic)
     for (int grid_index = 0; grid_index < this->nbxx; grid_index++) {
         const int na_grid = this->gridt->how_many_atoms[grid_index];
         if (na_grid == 0) {