### Background
In some cases, the computational workload varies significantly across different grids. Setting the OpenMP strategy to 'dynamic' can effectively balance the computational load and improve computational efficiency. Below is a comparison of computation times for one iteration of 'cal_gint_vlocal' before and after making these changes in several test cases:
||before|after|
|---|---|---|
|001_C2H6O|0.14s|0.08s|
|002_4MoS2|0.30s|0.13s|
|003_12Pt111|0.63s|0.27s|