@mohanchen 2025-02-13
The PR largely enhances the grid integration efficiency when dealing with vacuum (molecules, surfaces, etc.) Excellent!